Refacto:  Changed the naming of TaskProps to Task, Typed the handleChanged and handleSumbit functions and exported the types and queries in separated files
